+++ src/lib/servoarray.cpp
@@ -36,7 +36,7 @@ void ServoArray::set(std::uint8_t index, double rad) {
   this->values[index] = rad;
 
   const auto pulse = (rad + M_PI / 2) * (this->max_pulse - this->min_pulse) / M_PI + this->min_pulse;
-  this->driver.set_pwm(index, 0, pulse);
+  this->driver.set_pwm(index, 0, static_cast<std::uint16_t>(pulse));
 }
 
 double ServoArray::get(std::uint8_t index) {
